# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/formation": {
            "get": {
                "summary": "Retrieve a list of all formatmions",
                "operationId": "2f03c8a93255759863621022ff230bd5",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of formatmions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new formation",
                "operationId": "ce8438b5c676494a2d44637b38f17cc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Formation created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/formation/{id}": {
            "get": {
                "summary": "Retrieve a one formation ",
                "operationId": "372fd5d37c0743e1ae554c55c0e94032",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the formation you want show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully a formation selected"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a formation by ID",
                "operationId": "40f1486d2fe4f552ee9d4edfcdcc70db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the formation to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation updated successfully"
                    },
                    "404": {
                        "description": "Formation not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a formation",
                "operationId": "eaa087b9540372d06d38414bf1de8d61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the formation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "formation deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "formation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "summary": "Loged a new user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email (unique)",
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "description": "User's first name",
                                        "type": "string"
                                    },
                                    "residence": {
                                        "description": "User's residence",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User's email (unique)",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "profilePicture": {
                                        "description": "URL or base64-encoded image of the user's profile picture",
                                        "type": "string"
                                    },
                                    "levelOfStudy": {
                                        "description": "User's level of study",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}